4个任务：1 lcd任务 2.wave任务 3.模拟motor任务 4led显示系统正在运行。

+++ wave-iic/ks103/ks103.h
@@ -6,7 +6,7 @@
 
 u8 KS103_ReadOneByte(u8 address, u8 reg);
 void KS103_WriteOneByte(u8 address,u8 reg,u8 command);
-u16 Wave_test(void);
+int Wave_test(void);
 u16 Get_Range(u8 addr,int index);
 int Get_Barrier(u8 addr,int index);
 #endif
